+++ e2e/tests/constants.js
@@ -8,7 +8,12 @@ export const fixtures = [
   'facebook-video',
   'gif',
   'giphy',
+  'html',
+  'gallery-layouts',
+  'old-image-format',
 ];
+
+export const fixturesToTestOnSeo = ['images'];
 
 export const DEFAULT_DESKTOP_BROWSERS = [
   { width: 1440, height: 900, name: 'chrome' },