+++ examples/storybook/stories/index.js
@@ -1,7 +1,6 @@
 import './styles.global.scss';
 
 require('./Intro/');
-require('./Editor/');
-require('./Viewer/');
+require('./Plugins/');
 require('./Wrapper/');
 require('./TestCase/');