+++ packages/plugin-file-upload/web/src/index.js
@@ -2,3 +2,4 @@ export { createFileUploadPlugin } from './createFileUploadPlugin';
 export { default as FileUploadViewer } from './file-upload-viewer';
 export { FILE_UPLOAD_TYPE } from './types';
 export { typeMapper as fileUploadTypeMapper } from './typeMapper';
+export { pluginFileUpload } from './editor';