+++ packages/plugin-headers-markdown/web/src/defaults.js
@@ -1,0 +1,6 @@
+export const DEFAULTS = {
+  config: {},
+  configViewer: {
+    hideMarkdown: true,
+  },
+};