+++ packages/plugin-headers-markdown/web/src/index.js
@@ -3,3 +3,4 @@ export {
   createHeadersMarkdownDecorator,
 } from './createHeadersMarkdownPlugin';
 export const HEADERS_MARKDOWN_TYPE = 'wix-draft-plugin-headers-markdown';
+export { pluginHeadersMarkdown } from './editor';