+++ packages/wrapper/web/src/PluginsStrategy/index.js
@@ -1,2 +1,0 @@
-import pluginsStrategyProvider from './PluginsStrategyProvider';
-export default pluginsStrategyProvider;